requirements.txt
amqp==5.2.0
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==24.2.0
bcrypt==3.2.0
billiard==4.2.1
blinker==1.4
Brlapi==0.8.3
celery==5.4.0
certifi==2024.8.30
chardet==4.0.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
cron-descriptor==1.4.5
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
duplicity==0.8.21
exceptiongroup==1.2.2
fasteners==0.14.1
filelock==3.13.1
future==0.18.2
gyp==0.1
h11==0.14.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
keyring==23.5.0
kombu==5.4.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
outcome==1.3.0.post0
packaging==24.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
pluggy==1.5.0
prompt_toolkit==3.0.48
protobuf==3.12.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
pytest==8.3.3
pytest-django==4.9.0
python-apt==2.4.0+ubuntu4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==5.2.0
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
systemd-python==234
tomli==2.0.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.1.0
virtualenv==20.25.0
wadllib==1.3.6
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
amqp==5.2.0
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==24.2.0
bcrypt==3.2.0
billiard==4.2.1
blinker==1.4
Brlapi==0.8.3
celery==5.4.0
certifi==2024.8.30
chardet==4.0.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
cron-descriptor==1.4.5
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
duplicity==0.8.21
exceptiongroup==1.2.2
fasteners==0.14.1
filelock==3.13.1
future==0.18.2
gyp==0.1
h11==0.14.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
keyring==23.5.0
kombu==5.4.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
outcome==1.3.0.post0
packaging==24.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
pluggy==1.5.0
prompt_toolkit==3.0.48
protobuf==3.12.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
pytest==8.3.3
pytest-django==4.9.0
python-apt==2.4.0+ubuntu4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==5.2.0
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
systemd-python==234
tomli==2.0.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.1.0
virtualenv==20.25.0
wadllib==1.3.6
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
